+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Features
+
+- Add automatic UPnP MediaRenderer output discovery and playback.
+
 ## 1.1 (2026-05-30)
 
 ### Features